Implement LargeLocalizedStrings analysis for iOS (#137)
* Create outline for LargeLocalizedStrings analaysis for iOS

* Implement Localized Strings Insight.

This will trigger when the total size of the localized strings founds
exceeds 100KB to match the previous logic.

Happy to change the threshold or implement this in some other way.
For context, this is the action that should be performed to minimize the
strings: https://docs.emergetools.com/docs/minify-localized-strings