flush() when uploading
This changes two things:
- Adds a flush() prior to uploading. As with the previous issue with
  downloading it's possible we get into a bad state by reading from a
  file when the file has been written to but not flushed to disk.
- Re-add the duration to the analyzers.

In a previous patch I ended up removing the analyze_duration from the
reported results. I did this because it seemed misleading since:
a) it did not include artifact parsing which is a cost shared between
   size and distro
b) it did not include the duration of retries (if any)

However one theory is that this exacerbates problems with corrupt
uploads due to the size results being too deterministic (e.g. on a
retry you attempt to upload an identical file again but since it has
the same hash you end up seeing some of the state from the previous
upload).